Init UI using Code Generate.

+++ Forms/BookListForm.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
 
@@ -7,12 +8,45 @@ namespace BookstoreManagement
     public partial class BookListForm : Form
     {
         private BookstoreDataDataContext db;
+        private DataGridView dgvBooks;
+        private Button btnAdd, btnEdit, btnDelete;
 
         public BookListForm()
         {
-            InitializeComponent();
+            this.Text = "Book List";
+            this.Size = new Size(900, 500);
+            this.StartPosition = FormStartPosition.CenterScreen;
+
             db = new BookstoreDataDataContext();
+            InitializeControls();
             LoadBooks();
+        }
+
+        private void InitializeControls()
+        {
+            dgvBooks = new DataGridView
+            {
+                Top = 20,
+                Left = 20,
+                Width = 840,
+                Height = 350,
+                ReadOnly = true,
+                SelectionMode = DataGridViewSelectionMode.FullRowSelect,
+                AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.Fill
+            };
+            this.Controls.Add(dgvBooks);
+
+            btnAdd = new Button { Text = "Add", Top = 390, Left = 20, Width = 100 };
+            btnEdit = new Button { Text = "Edit", Top = 390, Left = 140, Width = 100 };
+            btnDelete = new Button { Text = "Delete", Top = 390, Left = 260, Width = 100 };
+
+            btnAdd.Click += btnAdd_Click;
+            btnEdit.Click += btnEdit_Click;
+            btnDelete.Click += btnDelete_Click;
+
+            this.Controls.Add(btnAdd);
+            this.Controls.Add(btnEdit);
+            this.Controls.Add(btnDelete);
         }
 
         private void LoadBooks()
@@ -22,12 +56,12 @@ namespace BookstoreManagement
                 b.Id,
                 b.Name,
                 Author = b.AuthorFullName,
-                b.Genre.Name,
-                b.PublishingHouse,
-                b.Pages,
-                b.DatePublished,
+                Genre = b.Genre.Name,
+                PublishingHouse = b.PublishingHouse,
+                Pages = b.Pages,
+                DatePublished = b.DatePublished,
                 b.SalePrice,
-                b.IsSequel
+                IsSequel = (bool)b.IsSequel ? "Yes" : "No"
             }).ToList();
 
             dgvBooks.DataSource = books;
@@ -35,11 +69,9 @@ namespace BookstoreManagement
 
         private void btnAdd_Click(object sender, EventArgs e)
         {
-            var form = new BookForm();
-            if (form.ShowDialog() == DialogResult.OK)
-            {
-                LoadBooks();
-            }
+            var form = new Forms.BookForm();
+            form.ShowDialog();
+            LoadBooks();
         }
 
         private void btnEdit_Click(object sender, EventArgs e)
@@ -51,11 +83,9 @@ namespace BookstoreManagement
 
                 if (book != null)
                 {
-                    var form = new BookForm(book);
-                    if (form.ShowDialog() == DialogResult.OK)
-                    {
-                        LoadBooks();
-                    }
+                    var form = new Forms.BookForm(book);
+                    form.ShowDialog();
+                    LoadBooks();
                 }
             }
         }
@@ -69,9 +99,13 @@ namespace BookstoreManagement
 
                 if (book != null)
                 {
-                    db.Books.DeleteOnSubmit(book);
-                    db.SubmitChanges();
-                    LoadBooks();
+                    var result = MessageBox.Show($"Are you sure you want to delete \"{book.Name}\"?", "Confirm Delete", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                    if (result == DialogResult.Yes)
+                    {
+                        db.Books.DeleteOnSubmit(book);
+                        db.SubmitChanges();
+                        LoadBooks();
+                    }
                 }
             }
         }
